Задание 5. Коммит с модулями.
